+++ RELEASE-NOTES.md
@@ -36,6 +36,7 @@
 * #371 ECharts: support "itemStyle" for "boxplot" series
 * #382 $bool(..).cumSum()
 * #413 ByteSource: generic data source for loaders
+* #414 HTTP connector with ByteSource
 * #415 CsvLoader to support an explicit encoding
 * #418 Exception printing null columns